Scrap builder, use simple config instead. Most stuff is optional anyway.
Using the builder caused RAII problems. The curl handle was being cleaned
up during instantiation of the actual curl wrapper, which destroyed a
builder instance. This necessitated that the cleanup code be moved to the
curl wrapper's destructor, which in turn needed access to the curl handle.
Now the handle is a member of the wrapper instead and never gets passed
around, so only gets cleaned up when the wrapper is destroyed.