Adds SSL certificate for GitHub and declares anonymous functions.
This is a huge changeset, also a consequence of earlier mentioned whoopsie.
Firstly it introduces an in-memory SSL certificate so curl can connect to
GitHub. Secondly it adds a function, curlcheck(), which evaluates a
CURLcode and throws an exception if the return value is not CURLE_OK.
Thirdly it does a bit of stylistic cleanup.

The reason this wasn't broken into smaller changesets is that add -p
couldn't automatically split the hunks acceptably and I'm too lazy to do it
manually.