Add support for multiple certificates.
GitHub's release downloads are hosted by Amazon, which uses a different
certificate from GitHub's. However, you still need to get past GitHub
to get to Amazon, so both certificates are necessary. This change
firstly adds the required Amazon certificate, secondly generalizes the
sslctxfunction to work on an array of certificates; not a vector,
because VS2010 doesn't support initializer lists.