Use OS layer TLS verification
Maintaining TLS certificates manually is painful and error-prone, as
evidenced by the expired Verisign certificate that now breaks updates.
Remove the hard-coded certificate validation and rely on the OS layer;
for GNU/Linux, this happens automatically insofar as the user has
installed certificates, but for Windows it requires libcurl with WinSSL:

    nmake /f Makefile.vc mode=static ENABLE_SSPI=yes ENABLE_IDN=no \
        ENABLE_WINSSL=yes DEBUG=no MACHINE=x86 GEN_PDB=no ENABLE_IPV6=yes

This completely eliminates the OpenSSL TLS dependency in the curl
wrapper, so those libraries can be removed. The only remaining
dependency on OpenSSL is checksum calculation opening up for static
linking.